apply fix to arithmetic question
