+++ packages/datatrak-web/src/features/Questions/index.ts
@@ -1,15 +1,14 @@
-export { TextQuestion } from './TextQuestion';
-export { RadioQuestion } from './RadioQuestion';
-export { BinaryQuestion } from './BinaryQuestion';
-export { DateQuestion } from './DateQuestion';
-export { InstructionQuestion } from './InstructionQuestion';
-export { CheckboxQuestion } from './CheckboxQuestion';
-export { DateTimeQuestion } from './DateTimeQuestion';
-export { GeolocateQuestion } from './GeolocateQuestion';
-export { EntityQuestion } from './EntityQuestion';
 export { AutocompleteQuestion } from './AutocompleteQuestion';
-export { ReadOnlyQuestion, CodeGeneratorQuestion } from './ReadOnlyQuestion';
-export { ArithmeticQuestion } from './ArithmeticQuestion';
-export { PhotoQuestion } from './PhotoQuestion';
+export { BinaryQuestion } from './BinaryQuestion';
+export { CheckboxQuestion } from './CheckboxQuestion';
+export { DateQuestion } from './DateQuestion';
+export { DateTimeQuestion } from './DateTimeQuestion';
+export { EntityQuestion } from './EntityQuestion';
 export { FileQuestion } from './FileQuestion';
+export { GeolocateQuestion } from './GeolocateQuestion';
+export { InstructionQuestion } from './InstructionQuestion';
+export { PhotoQuestion } from './PhotoQuestion';
+export { RadioQuestion } from './RadioQuestion';
+export { ArithmeticQuestion, CodeGeneratorQuestion, ReadOnlyQuestion } from './ReadOnlyQuestion';
+export { TextQuestion } from './TextQuestion';
 export { UserQuestion } from './UserQuestion';
